notification code moved to utils file
